Work over main.nf
Work over the language: Replace separate splittings of channels with multiple channel declarations in the process.
Remove confusing declarations of file-output-namesin the script-block --> use it in the 'saveAs' command.
Go over the help-function to include changes made over the past half
year and remove parameters not implemented in the pipeline (e.g.
--dedup)