Replace hardcoded Mappings to wanted species with File
One can now add a --specmap TSV_FILE flag to the nextflow command.
the format of the file: 'Family<tab>Species_name,Species_name'.
Assigned "best species" by Kraken are overwritten by the species in
that file and mapped to. If assigned more than one species, the
best one of these is still picked after the deduplication process
for bedfiltering.