Add optional deamination analysis to end of pipeline
use the tag --analyze to add an anlysis step to the end of the
pipeline. The resulting table is found in stats/rg.deamination_stats.tsv
The table contains for the first and last three basepairs the normal deamination
frequencies (C to T substitutions) and the conditional substitutions as well as
the respective 95% confidence intervals. See the README for more information