Filter krakenuniq-reports
The CL header in the krakenuniq-reports (after running in singularity)
contain extra lines with junk-information

That is due to the way krakenuniq fetches the current command
(ps -o args=PID), which returns A LOT of lines executed
within a singularity container.

This workaround just filters the report again after being created.

Not nice, but until krakenuniq is not fixed (and we dont switch to conda)
there is not really another way.