Write all reports at the end to handle --onlyfixed runs
to not overwrite the files in the `stats` directory with
the 'fixed' file only, write all reports
at the end of the run, after merging the old report with the
new meta-data. Thus, all the entries are kept!

This would even allow the processing of serveral split-dirs
in the same directory if RG-names are not conflicting.